Struct google_dfareporting3d2::api::Ad

pub struct Ad {
    pub account_id: Option<String>,
    pub active: Option<bool>,
    pub advertiser_id: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub archived: Option<bool>,
    pub audience_segment_id: Option<String>,
    pub campaign_id: Option<String>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub click_through_url: Option<ClickThroughUrl>,
    pub click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>,
    pub comments: Option<String>,
    pub compatibility: Option<String>,
    pub create_info: Option<LastModifiedInfo>,
    pub creative_group_assignments: Option<Vec<CreativeGroupAssignment>>,
    pub creative_rotation: Option<CreativeRotation>,
    pub day_part_targeting: Option<DayPartTargeting>,
    pub default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>,
    pub delivery_schedule: Option<DeliverySchedule>,
    pub dynamic_click_tracker: Option<bool>,
    pub end_time: Option<String>,
    pub event_tag_overrides: Option<Vec<EventTagOverride>>,
    pub geo_targeting: Option<GeoTargeting>,
    pub id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub key_value_targeting_expression: Option<KeyValueTargetingExpression>,
    pub kind: Option<String>,
    pub language_targeting: Option<LanguageTargeting>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub name: Option<String>,
    pub placement_assignments: Option<Vec<PlacementAssignment>>,
    pub remarketing_list_expression: Option<ListTargetingExpression>,
    pub size: Option<Size>,
    pub ssl_compliant: Option<bool>,
    pub ssl_required: Option<bool>,
    pub start_time: Option<String>,
    pub subaccount_id: Option<String>,
    pub targeting_template_id: Option<String>,
    pub technology_targeting: Option<TechnologyTargeting>,
    pub type_: Option<String>,
}

Contains properties of a Campaign Manager ad.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get ads (response)

insert ads (request|response)

list ads (none)

patch ads (request|response)

update ads (request|response)

Fields
account_id: Option<String>

Account ID of this ad. This is a read-only field that can be left blank.

active: Option<bool>

Whether this ad is active. When true, archived must be false.

advertiser_id: Option<String>

Advertiser ID of this ad. This is a required field on insertion.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

archived: Option<bool>

Whether this ad is archived. When true, active must be false.

audience_segment_id: Option<String>

Audience segment ID that is being targeted for this ad. Applicable when type is AD_SERVING_STANDARD_AD.

campaign_id: Option<String>

Campaign ID of this ad. This is a required field on insertion.

campaign_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

click_through_url: Option<ClickThroughUrl>

Click-through URL for this ad. This is a required field on insertion. Applicable when type is AD_SERVING_CLICK_TRACKER.

click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>

Click-through URL suffix properties for this ad. Applies to the URL in the ad or (if overriding ad properties) the URL in the creative.

comments: Option<String>

Comments for this ad.

compatibility: Option<String>

Compatibility of this ad. Applicable when type is AD_SERVING_DEFAULT_AD. DISPLAY and DISPLAY_INTERSTITIAL refer to either rendering on desktop or on mobile devices or in mobile apps for regular or interstitial ads, respectively. APP and APP_INTERSTITIAL are only used for existing default ads. New mobile placements must be assigned DISPLAY or DISPLAY_INTERSTITIAL and default ads created for those placements will be limited to those compatibility types. IN_STREAM_VIDEO refers to rendering in-stream video ads developed with the VAST standard.

create_info: Option<LastModifiedInfo>

Information about the creation of this ad. This is a read-only field.

creative_group_assignments: Option<Vec<CreativeGroupAssignment>>

Creative group assignments for this ad. Applicable when type is AD_SERVING_CLICK_TRACKER. Only one assignment per creative group number is allowed for a maximum of two assignments.

creative_rotation: Option<CreativeRotation>

Creative rotation for this ad. Applicable when type is AD_SERVING_DEFAULT_AD, AD_SERVING_STANDARD_AD, or AD_SERVING_TRACKING. When type is AD_SERVING_DEFAULT_AD, this field should have exactly one creativeAssignment.

day_part_targeting: Option<DayPartTargeting>

Time and day targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>

Default click-through event tag properties for this ad.

delivery_schedule: Option<DeliverySchedule>

Delivery schedule information for this ad. Applicable when type is AD_SERVING_STANDARD_AD or AD_SERVING_TRACKING. This field along with subfields priority and impressionRatio are required on insertion when type is AD_SERVING_STANDARD_AD.

dynamic_click_tracker: Option<bool>

Whether this ad is a dynamic click tracker. Applicable when type is AD_SERVING_CLICK_TRACKER. This is a required field on insert, and is read-only after insert.

end_time: Option<String>

Date and time that this ad should stop serving. Must be later than the start time. This is a required field on insertion.

event_tag_overrides: Option<Vec<EventTagOverride>>

Event tag overrides for this ad.

geo_targeting: Option<GeoTargeting>

Geographical targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

id: Option<String>

ID of this ad. This is a read-only, auto-generated field.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this ad. This is a read-only, auto-generated field.

key_value_targeting_expression: Option<KeyValueTargetingExpression>

Key-value targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#ad".

language_targeting: Option<LanguageTargeting>

Language targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this ad. This is a read-only field.

name: Option<String>

Name of this ad. This is a required field and must be less than 256 characters long.

placement_assignments: Option<Vec<PlacementAssignment>>

Placement assignments for this ad.

remarketing_list_expression: Option<ListTargetingExpression>

Remarketing list targeting expression for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

size: Option<Size>

Size of this ad. Applicable when type is AD_SERVING_DEFAULT_AD.

ssl_compliant: Option<bool>

Whether this ad is ssl compliant. This is a read-only field that is auto-generated when the ad is inserted or updated.

ssl_required: Option<bool>

Whether this ad requires ssl. This is a read-only field that is auto-generated when the ad is inserted or updated.

start_time: Option<String>

Date and time that this ad should start serving. If creating an ad, this field must be a time in the future. This is a required field on insertion.

subaccount_id: Option<String>

Subaccount ID of this ad. This is a read-only field that can be left blank.

targeting_template_id: Option<String>

Targeting template ID, used to apply preconfigured targeting information to this ad. This cannot be set while any of dayPartTargeting, geoTargeting, keyValueTargetingExpression, languageTargeting, remarketingListExpression, or technologyTargeting are set. Applicable when type is AD_SERVING_STANDARD_AD.

technology_targeting: Option<TechnologyTargeting>

Technology platform targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

type_: Option<String>

Type of ad. This is a required field on insertion. Note that default ads (AD_SERVING_DEFAULT_AD) cannot be created directly (see Creative resource).