Struct google_dfareporting3d2::api::TargetingTemplateMethods

pub struct TargetingTemplateMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on targetingTemplate resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting3d2 as dfareporting3d2;
 
use std::default::Default;
use oauth2;
use dfareporting3d2::Dfareporting;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Dfareporting::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.targeting_templates();

Implementations

impl<'a, C> TargetingTemplateMethods<'a, C>

pub fn get(&self, profile_id: &str, id: &str) -> TargetingTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets one targeting template by ID.

Arguments

profileId - User profile ID associated with this request.

id - Targeting template ID.

pub fn insert(
&self,
request: TargetingTemplate,
profile_id: &str
) -> TargetingTemplateInsertCall<'a, C>

Create a builder to help you perform the following task:

Inserts a new targeting template.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

pub fn list(&self, profile_id: &str) -> TargetingTemplateListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a list of targeting templates, optionally filtered. This method supports paging.

Arguments

profileId - User profile ID associated with this request.

pub fn patch(
&self,
request: TargetingTemplate,
profile_id: &str,
id: &str
) -> TargetingTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing targeting template. This method supports patch semantics.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

id - Targeting template ID.

pub fn update(
&self,
request: TargetingTemplate,
profile_id: &str
) -> TargetingTemplateUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing targeting template.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.